Move downloadProfile method to ServiceInstance
